backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
google-auth
requests
